debug

+++ examples/run_nConv_pos.py
@@ -229,7 +229,7 @@ def train(args, train_dataset, model, tokenizer):
                 continue
 
             model.train()
-            batch = tuple(t.half().to(args.device) for t in batch)
+            batch = tuple(t.to(args.device) for t in batch)
             inputs = {"input_ids": batch[0], "attention_mask": batch[1], "labels": batch[3], "tag_ids":batch[4]}
             if args.model_type != "distilbert":
                 inputs["token_type_ids"] = (
